requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
greenlet==3.1.1
h11==0.14.0
html5tagger==1.3.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sanic==24.12.0
sanic-ext==23.12.0
sanic-jwt==1.8.0
sanic-routing==23.12.0
sanic-testing==24.6.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvloop==0.21.0
websockets==14.2
